[SymForce] Check in generated LCM files
Check in LCM files just like we do the generated sym files. This
greatly simplifies the setup.py build because we don't have to
install a package created from cmake, that is not an extension. We
should add a test to verify the generated files. I'm also sure we
can work around this in setup.py if we really don't want to do this.

Topic: symforce/pip_install